+++ NEWS.md
@@ -1,5 +1,7 @@
 # CHANGES IN xfun VERSION 0.44
 
+- Added a function `cache_exec()` to cache the execution of an expression either in memory or on disk. It is much more general and flexible than `cache_rds()`. For example, it supports custom reading/writing methods for cache files, and can load locally created variables in the expression while loading cache.
+
 - Exported the internal function `md5()` to calculate the MD5 checksums of R objects. The function is essentially a workaround for `tools::md5sum()` (see HenrikBengtsson/Wishlist-for-R#21).
 
 - `write_utf8()` returns the `con` argument (typically a file path) now. Previously, it returns `NULL`.